+++ eggs/Naaya/CHANGELOG.rst
@@ -3,6 +3,7 @@
 * Fix news and stories folder sort order [andredor]
 * Zip download link is no longer shown if there are no objects to download
   [dumitval]
+* OpenLayers map engine [moregale]
 
 2.12.5 (2011-10-11)
 -------------------